pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "isqa"
description = "A CLI utility for quality assurance of GitLab issues."
version = "0.5.0"
readme = "README.md"
authors = [
  {name = "vdmitriyev"}
]
maintainers = [
  {name = "vdmitriyev"}
]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "python-gitlab>=4.1.0",
    "python-dotenv>=1.0.0",
    "Jinja2>=3.1.6"
]

[project.scripts]
isqa = "isqa.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["isqa*"]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "pre-commit>=4.5.1",
]
